requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pdfkit==1.0.0
pydantic>=2.8.2
google-genai>=1.0.0
python-dotenv==1.2.1
